Rename robot service file
